+++ src/dev_tools.rs
@@ -4,7 +4,6 @@ use bevy::{
     dev_tools::states::log_transitions, diagnostic::FrameTimeDiagnosticsPlugin, prelude::*,
 };
 use bevy_inspector_egui::quick::WorldInspectorPlugin;
-use avian2d::prelude::PhysicsDebugPlugin;
 
 use crate::ui::screens::Screen;
 
@@ -17,7 +16,6 @@ impl Plugin for DebugPlugin {
         app.add_systems(Update, log_transitions::<Screen>);
         app.add_plugins((
             WorldInspectorPlugin::default(),
-            PhysicsDebugPlugin::default(),
         ));
     }
 }
