+++ src/game/mod.rs
@@ -4,14 +4,12 @@ use bevy::prelude::*;
 
 pub mod audio;
 pub mod camera;
-pub mod physics;
-pub mod spawn;
+pub mod oscilloscope;
 
 pub(super) fn plugin(app: &mut App) {
     app.add_plugins((
         camera::plugin,
         audio::plugin,
-        spawn::plugin,
-        physics::plugin,
+        oscilloscope::plugin,
     ));
 }
