+++ src/ui/screens/mod.rs
@@ -11,7 +11,6 @@ use bevy::prelude::*;
 pub(super) fn plugin(app: &mut App) {
     app.init_state::<Screen>();
     app.enable_state_scoped_entities::<Screen>();
-
     app.add_plugins((
         splash::plugin,
         loading::plugin,
