+++ src/ui/screens/playing.rs
@@ -3,7 +3,7 @@
 use bevy::prelude::*;
 
 use super::Screen;
-use crate::{game::spawn::*, ui::palette::OSCILLOSCOPE_SCREEN_COLOR};
+use crate::{game::oscilloscope, ui::palette::OSCILLOSCOPE_SCREEN_COLOR};
 
 pub(super) fn plugin(app: &mut App) {
     app.add_systems(OnEnter(Screen::Playing), enter_playing);
